Add in FrequencyFilteredNoiseMapCreatorTask support for MapCreator class

+++ test/resources/data/maps/filtered_noise_map/dependent.py
@@ -1,3 +1,2 @@
-def dependent(x: int, y: int, get_tile) -> int:
+def dependent(x: int, y: int, get_tile) -> float:
     return get_tile(x,y).base - 1
-
